requirements.txt

google-cloud-datastore==1.6.0
google-cloud-storage==1.9.0
h5py==2.7.1
Keras==2.1.6
numpy==1.14.3
pandas==0.22.0
PyYAML==3.12
scikit-learn==0.19.1
scipy==1.1.0
tensorboard==1.8.0
tensorflow-gpu==1.8.0
